"ads2_m13_a1_1: 0/8 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/m13/Radix Sorts/Solution.java
@@ -26,7 +26,7 @@ public class Solution {
 
         // read in the strings from standard input
         Scanner sc = new Scanner(System.in);
-        int numOfStrings = sc.nextInt();
+        double numOfStrings = sc.nextDouble();
         sc.nextLine();
         Quick3string quick3way = new Quick3string();
         String[] a = StdIn.readAllStrings();
